Inheritance Cube: Immunizing Reversal/Basic Reversal

This function creates an Immunizing Reversal construct, it has 5 alleles at 1 locus

W: Wild-type

H: Homing allele

E: Eraser allele

R: No-cost resistance allele

B: Detrimental resistance allele

eHWM: Rate of homing in males, W -> H transition

eEWM: Rate of homing in males, W -> E transition

eEHM: Rate of homing in males, H -> E transition

eHWF: Rate of homing in females, W -> H transition

eEWF: Rate of homing in females, W -> E transition

eEHF: Rate of homing in females, H -> E transition

rHWM: Generation of R alleles during homing in males, from H

bHWM: Generation of B alleles during homing in males, from H

rEWM: Generation of R alleles during homing in males, from E

bEWM: Generation of B alleles during homing in males, from E

rEHM: Generation of R alleles during homing in males, from E

bEHM: Generation of B alleles during homing in males, from E

rHWF: Generation of R alleles during homing in females, from H

bHWF: Generation of B alleles during homing in females, from H

rEWF: Generation of R alleles during homing in females, from E

bEWF: Generation of B alleles during homing in females, from E

rEHF: Generation of R alleles during homing in females, from E

bEHF: Generation of B alleles during homing in females, from E

Cube_ImmunizingReversalMF(eHWM = 1, eEWM = 1, eEHM = 1, eHWF = 1,
  eEWF = 1, eEHF = 1, rHWM = 0, bHWM = 0, rEWM = 0, bEWM = 0,
  rEHM = 0, bEHM = 0, rHWF = 0, bHWF = 0, rEWF = 0, bEWF = 0,
  rEHF = 0, bEHF = 0, eta = NULL, phi = NULL, omega = NULL,
  xiF = NULL, xiM = NULL, s = NULL)

Arguments

eHWM
Homing efficiency of drive allele in males

eEWM
Homing efficiency of IR allele into wild-type allele in males

eEHM
Homing efficiency of IR allele into homing allele in males

eHWF
Homing efficiency of drive allele into wild-type allele in females

eEWF
Homing efficiency of IR allele into wild-type allele in females

eEHF
Homing efficiency of IR allele into homing allele in females

rHWM
No-cost resistant allele generation rate in males, from H into W

bHWM
Detrimental resistant allele genration rate in males, from H into W

rEWM
No-cost resistant allele generation rate in males, from E into W

bEWM
Detrimental resistant allele generation rate in males, from E into W

rEHM
No-cost resistant allele generation rate in males, from E into H

bEHM
Detrimental resistant allele generation rate in males, from E into H

rHWF
No-cost resistant allele generation rate in females, from H into W

bHWF
Detrimental resistant allele genration rate in females, from H into W

rEWF
No-cost resistant allele generation rate in females, from E into W

bEWF
Detrimental resistant allele generation rate in females, from E into W

rEHF
No-cost resistant allele generation rate in females, from E into H

bEHF
Detrimental resistant allele generation rate in females, from E into H

eta
Genotype-specific mating fitness

phi
Genotype-specific sex ratio at emergence

omega
Genotype-specific multiplicative modifier of adult mortality

xiF
Genotype-specific female pupatory success

xiM
Genotype-specific male pupatory success

s
Genotype-specific fractional reduction(increase) in fertility

Value

Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele, and all genotype-specific parameters.

Details

This is the general form for an immunizing reversal drive. If the eEWM and eEWF are 0, and the other (r/b)EW(M/F) parameters are 0, then this simplifies to a basic reversal drive.This drive handles different male and female homing rates, signifying differential expression from an autosome.