Inheritance Cube: ECHACR

This function creates an ECHACR construct, it has 5 alleles at the first locus and 4 alleles at the second.

W: Wild-type

H: Homing allele

E: Eraser allele

R: No-cost resistance allele

B: Detrimental resistance allele

eH: Rate of homing from H, W -> H transition

eE: Rate of homing from E, H -> E transition

pRH: Generation of R alleles during homing, from H

pBH: Generation of B alleles during homing, from H

pRE: Generation of R alleles during homing, from E

pBE: Generation of B alleles during homing, from E

Cube_eCHACR(eH = 1, eW = 1, eC = 1, eCH = 1, eCW = 1, eCC = 1,
  pRH = 0, pBH = 0, pRW = 0, pBW = 0, pRC = 0, pBC = 0, pRCH = 0,
  pBCH = 0, pRCW = 0, pBCW = 0, pRCC = 0, pBCC = 0, d = 0,
  eta = NULL, phi = NULL, omega = NULL, xiF = NULL, xiM = NULL,
  s = NULL)

Arguments

eH
Homing efficiency of drive allele at locus 1

eW
Homing efficiency of ECHACR element at locus 2

eC
Homing efficiency of ECHACR element into homing allele at locus 1

eCH
Homing efficiency of drive element in presence of ECHACR element

eCW
Homing efficiency of ECHACR element in presence of 1 homing allele into new allele at locus 1

eCC
Homing efficiency of ECHACR element in presence of 1 homing allele and 1 not at locus 1

pRH
No-cost resistant allele generation rate

pBH
Detrimental resistant allele generation rate

pRW
No-cost resistante allele from eW

pBW
Detrimental resistant allele from eW

pRC
No-cost resistance allele from eC

pBC
Detrimental resistant allele from eC

pRCH
No-cost resistance allele from eCH

pBCH
Detrimental resistant allele from eCH

pRCW
No-cost resistant allele from eCW

pBCW
Detrimental resistant allele from eCW

pRCC
No-cost resistant allele from eCC

pBCC
Detrimental resistant allele from eCC

d
Background mutation rate from W,H, and E into R allele

eta
Genotype-specific mating fitness

phi
Genotype-specific sex ratio at emergence

omega
Genotype-specific multiplicative modifier of adult mortality

xiF
Genotype-specific female pupatory success

xiM
Genotype-specific male pupatory success

s
Genotype-specific fractional reduction(increase) in fertility

Value

Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele, and all genotype-specific parameters.