Network Parameters

Generate parameters for simulation on a Network. Parameters average generation time \(g\), population growth rate \(R_{m}\), aquatic mortality \(\mu_{Aq}\), and aquatic survival \(\theta_{Aq}\) are shared between patches and calculated by calcAverageGenerationTime, calcPopulationGrowthRate, calcLarvalStageMortalityRate, and calcAquaticStagesSurvivalProbability.
Patch-specific parameters \(\alpha\) and \(L_{eq}\) are calculated for each patch by calcDensityDependentDeathRate and calcLarvalPopEquilibrium.

Network.Parameters(nPatch, simTime, parallel = FALSE, moveVar = 1000,
  tEgg = 1L, tLarva = 14L, tPupa = 1L, beta = 32, muAd = 0.123,
  popGrowth = 1.096, AdPopEQ, runID = 1L)

Arguments

nPatch
number of Patch

simTime
maximum time to run simulation

parallel
append process id (see link[base]{Sys.getpid}) to output files for running in parallel

moveVar
variance of stochastic movement (not used in diffusion model of migration). It affects the concentration of probability in the Dirchlet simplex, small values lead to high variance and large values lead to low variance.

tEgg
length of egg stage

tLarva
length of larval instar stage

tPupa
length of pupal stage

beta
female egg batch size of wild-type

muAd
wild-type daily adult mortality (1/muAd is average wild-type lifespan)

popGrowth
daily population growth rate (used to calculate equilibrium)

AdPopEQ
vector of adult population size at equilibrium

runID
begin counting runs with this set of parameters from this value