Patch Class Definition

A Patch is a single well-mixed population that is the smallest unit of simulation for MGDrivE.

Format

An R6Class generator object

Constructor

patchID: integer ID of this patch

genotypesID: character vector of genotypes

timeAq: integer vector of length 3 specifying the length of each aquatic stage

numPatches: integer, total number of patches in this simulation

adultEQ: integer, total adult population in this patch for the duration of the simulation

larvalEQ: integer, total larval population in this patch for the duration of the simulation

muAq: double vector, length 3, daily death rate for each aquatic stage

alpha: double, density-dependent centering parameter, see parameterizeMGDrivE

adultRatioF: named double vector, distribution of adult female genotypes, see parameterizeMGDrivE

adultRatioM: named double vector, distribution of adult male genotypes, see parameterizeMGDrivE

larvalRatio: named double vector, distribution of all aquatic genotypes, see parameterizeMGDrivE

eggReleases: egg release schedule for this patch, see basicRepeatedReleases

maleReleases: male release schedule for this patch, see basicRepeatedReleases

femaleReleases: female release schedule for this patch, see basicRepeatedReleases

matedFemaleReleases: mated females release schedule for this patch, see basicRepeatedReleases

Methods

set_NetworkPointer: see set_NetworkPointer_Patch

get_femalePopulation: see get_femalePop_Patch

get_malePopulation: see get_malePop_Patch

initialPopulation: see set_initialPopulation_Patch

setPopulation: see set_population_deterministic_Patch or set_population_stochastic_Patch

reset: see reset_Patch

oneDay_initOutput: see oneDay_initOutput_Patch

oneDay_writeOutput: see oneDay_writeOutput_Patch

oneDay_migrationIn: see oneDay_migrationIn_Patch

oneDay_PopDynamics: see oneDay_PopDynamics_Patch

oneDay_adultD: see oneDay_adultDeath_deterministic_Patch or oneDay_adultDeath_stochastic_Patch

oneDay_pupaDM: see oneDay_pupaDM_deterministic_Patch or oneDay_pupaDM_stochastic_Patch

oneDay_larvaDM: see oneDay_larvaDM_deterministic_Patch or oneDay_larvaDM_stochastic_Patch

oneDay_eggDM: see oneDay_eggDM_deterministic_Patch or oneDay_eggDM_stochastic_Patch

oneDay_pupation: see oneDay_pupation_deterministic_Patch or oneDay_pupation_stochastic_Patch

oneDay_releases: see oneDay_releases_Patch

oneDay_releaseEggs: see oneDay_eggReleases_Patch

oneDay_mating: see oneDay_mating_deterministic_Patch or oneDay_mating_stochastic_Patch

oneDay_layEggs: see oneDay_oviposit_deterministic_Patch or oneDay_oviposit_stochastic_Patch

Fields

patchID: integer ID of this patch

popAquatic: matrix, nGenotype x sum(timeAquatic), holding all eggs, larva, and pupa

popMale: vector, nGenotype x 1, holds adult males

popFemale: matrix, nGenotype x nGenotype, holds mated adult females

popHolder: vector, nGenotype x 1, temporary population storage

popPupSex: vector, nGenotype x 1, used in stochastic pupation as another temporary population

popUnmated: vector, nGenotype x 1, holds unmated females

popAquatict0: matrix, nGenotype x sum(timeAquatic), holding all eggs, larva, and pupa for reset, see reset_Patch

popMalet0: vector, nGenotype x 1, holds adult males for reset see reset_Patch

popFemalet0: matrix, nGenotype x nGenotype, holds mated adult females for reset see reset_Patch

eggReleases: list of egg releases for this patch. See oneDay_eggReleases_Patch

maleReleases: list of adult male releases for this patch. See oneDay_releases_Patch

femaleReleases: list of adult female releases for this patch. See oneDay_releases_Patch

matedFemaleReleases: list of mated adult female releases for this patch. See oneDay_releases_Patch

NetworkPointer: a reference to enclosing Network