Aggregate Female Output by Genotype

Aggregate over male mate genotype to convert female matrix output into vector output.

Usage

aggregateFemales(
  readDir,
  writeDir = NULL,
  genotypes,
  remFile = TRUE,
  verbose = TRUE
)

Arguments

readDir

Directory to read input from

writeDir

Directory to write output to. Default is readDir

genotypes

Character vector of possible genotypes; found in driveCube$genotypesID

remFile

Boolean flag to remove original (unaggregated) file

verbose

Chatty? Default is TRUE

Examples

if (FALSE) { # \dontrun{
# This example assumes user has already run MGDrivE and generated output.
#  This also assumes that the user has already split output by patch.
# See vignette for complete example.

# set read/write directory
fPath <- "path/to/data/containing/folder"

# Need genotypes from the cube run in the simulation
#  This is dependent on the simulation run
#  Using Mendelian cube for this example
cube <- cubeMendelian()

# no return value from function
aggregateFemales(readDir= fPath, writeDir = NULL, genotypes = cube$genotypesID,
                 remFile = TRUE)
} # }