Inheritance Cube: Split-Drive for Andrea Smidler

This is a modified split-drive construct for split-suppression drives. This is an autosomal, 2-locus drive. The first locus contains the Cas9 allele and the second locus has the gRNA construct. Cas9 efficacy in each sex is dependent upon parent of inheritance. This construct has 3 alleles at the first locus and 4 alleles at the second.

Locus 1

W: Wild-type

P: Paternal Cas9

M: Maternal Cas9

Locus 2

W: Wild-type

G: gRNAs

R: Resistant 1

B: Resistant 2

Usage

cubeASmidler(
  cMM = 0,
  chMM = 0,
  crMM = 0,
  cPM = 0,
  chPM = 0,
  crPM = 0,
  cMF = 0,
  chMF = 0,
  crMF = 0,
  cPF = 0,
  chPF = 0,
  crPF = 0,
  crF = 0.5,
  crM = 0.5,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)

Arguments

cMM

Cutting efficacy of maternally-inherited Cas9 in males

chMM

Homing efficacy of maternally-inherited Cas9 in males

crMM

Resistance rate of maternally-inherited Cas9 in males

cPM

Cutting efficacy of paternally-inherited Cas9 in males

chPM

Homing efficacy of paternally-inherited Cas9 in males

crPM

Resistance rate of paternally-inherited Cas9 in males

cMF

Cutting efficacy of maternally-inherited Cas9 in females

chMF

Homing efficacy of maternally-inherited Cas9 in females

crMF

Resistance rate of maternally-inherited Cas9 in females

cPF

Cutting efficacy of paternally-inherited Cas9 in females

chPF

Homing efficacy of paternally-inherited Cas9 in females

crPF

Resistance rate of paternally-inherited Cas9 in females

crF

Female crossover rate, 0 is completely linked, 0.5 is unlinked, 1.0 is complete divergence

crM

Male crossover rate, 0 is completely linked, 0.5 is unlinked, 1.0 is complete divergence

eta

Genotype-specific mating fitness

phi

Genotype-specific sex ratio at emergence

omega

Genotype-specific multiplicative modifier of adult mortality

xiF

Genotype-specific female pupatory success

xiM

Genotype-specific male pupatory success

s

Genotype-specific fractional reduction(increase) in fertility

Value

Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele, and all genotype-specific parameters.