Inheritance Cube: 3-Piece Allele Sail

A generalized implementation of the Allele Sail (doi:10.1038/s41467-024-50992-9 ) idea.

Usage

cubeAlleleSail(
  cMM = 0,
  crMM = 0,
  cPM = 0,
  crPM = 0,
  cMF = 0,
  crMF = 0,
  cPF = 0,
  crPF = 0,
  dMW = 0,
  dMrW = 0,
  dPW = 0,
  dPrW = 0,
  crF12 = 0.5,
  crM12 = 0.5,
  crF23 = 0.5,
  crM23 = 0.5,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)

Arguments

cMM

Cutting efficacy of maternally-inherited Cas9 in males

crMM

Resistance rate of maternally-inherited Cas9 in males

cPM

Cutting efficacy of paternally-inherited Cas9 in males

crPM

Resistance rate of paternally-inherited Cas9 in males

cMF

Cutting efficacy of maternally-inherited Cas9 in females

crMF

Resistance rate of maternally-inherited Cas9 in females

cPF

Cutting efficacy of paternally-inherited Cas9 in females

crPF

Resistance rate of paternally-inherited Cas9 in females

dMW

Female deposition cutting rate, maternal Cas9

dMrW

Female deposition functional resistance rate, maternal Cas9

dPW

Female deposition (HH) cutting rate, paternal Cas9

dPrW

Female deposition (HH) functional resistance rate, paternal Cas9

crF12

Female crossover rate between loci 1 and 2, 0 is completely linked, 0.5 is unlinked, 1.0 is complete divergence

crM12

Male crossover rate between loci 1 and 2, 0 is completely linked, 0.5 is unlinked, 1.0 is complete divergence

crF23

Female crossover rate between loci 2 and 3, 0 is completely linked, 0.5 is unlinked, 1.0 is complete divergence

crM23

Male crossover rate between loci 2 and 3, 0 is completely linked, 0.5 is unlinked, 1.0 is complete divergence

eta

Genotype-specific mating fitness

phi

Genotype-specific sex ratio at emergence

omega

Genotype-specific multiplicative modifier of adult mortality

xiF

Genotype-specific female pupatory success

xiM

Genotype-specific male pupatory success

s

Genotype-specific fractional reduction(increase) in fertility

Value

Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele, and all genotype-specific parameters.

Details

This is an autosomal, 3-locus system. The first locus contains the Cas9 allele, the second locus carries the gRNA, and the third locus is the target. All loci can be linked/unlinked to the locus before it (so, 1 to 2 or 2 to 3). Cas9 efficacy due to provenance (mother vs father) is included.

This construct is very similar to our 2-locus Cleave and Rescue design for Oberhofer et. al.(doi:10.1101/2020.07.09.196253 ).

This construct has 3 alleles at the first locus, 2 alleles at the second locus, and 3 alleles at the third locus.

Locus 1

W: Wild-type

P: Paternal Cas9

M: Maternal Cas9

Locus 2

W: Wild-type

G: gRNAs

Locus 3

W: Wild-type

R: Resistant 1

B: Resistant 2

Female deposition is implemented incorrectly. Right now, it is performed on male alleles prior to zygote formation - it should happen post-zygote formation. Since this construct doesn't have HDR, this should be fine.