Inheritance Cube: CLEAVR - Cleave and Rescue

This is a novel cube from the Akbari lab. There are 2 loci: the first is a female fertility locus (e.g. doubleSex), where the Cas, gRNAs, and a recoded essential gene go. This locus is inherited in a Mendelian fashion, but is also targeted for destruction by the homing allele. The second locus involves an essential gene, for both males and females, and this is the target of the gRNAs at the first locus. No homing is performed, it is simply destroyed. There is different cutting rates in males and females, with no possibility for a rescuing resistant allele. Females homozygous for the H or B alleles at locus 1 are viable but infertile, while males are unaffected. All animals homozygous at locus two must contain the recoded copy at locus 1 to be viable. This corresponds to the homing construct being X-linked

cubeCLEAVRX(cM1 = 1, cM2 = 1, cF1 = 1, cF2 = 1, eta = NULL,
  phi = NULL, omega = NULL, xiF = NULL, xiM = NULL, s = NULL)

Arguments

cM1
Male cutting rate at first locus IS NOT USED IN THIS CUBE

cM2
Male cutting rate at second locus

cF1
Female cutting rate at first locus

cF2
Female cutting rate at second locus

eta
Genotype-specific mating fitness

phi
Genotype-specific sex ratio at emergence

omega
Genotype-specific multiplicative modifier of adult mortality

xiF
Genotype-specific female pupatory success

xiM
Genotype-specific male pupatory success

s
Genotype-specific fractional reduction(increase) in fertility

Value

Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele, and all genotype-specific parameters.