Inheritance Cube: 2-Locus ClvR (Cleave and Rescue)

Based on the Cleave-and-Rescue system of Oberhofer, this is a 3-locus Cas9-based toxin-antidote system. The first locus carries the Cas9, the second locus carries the gRNAs, and a recoded copy of an essential gene. The third locus is the targeted essential gene. This gene can be completely haplosufficient (hSuf = 1) or completely haploinsufficient (hSuf = 0). It is assumed that having 2 copies of the gene (be it wild-type at the second locus or recoded at the first) confers complete viability. Additionally, loci 1 and 2 can be linked, given crM and crF, imitating the original 2-locus ClvR system. For this construct, the first locus will have 2 alleles, the second will have 2 alleles, and the third will have 3 alleles:

Locus 1

W: Wild-type

C: Cas9

Locus 2

W: Wild-type

G: gRNAs and recoded essential gene

Locus 3

W: Wild-type

R: Functional resistant

B: Non-functional resistant

Usage

cubeClvR2(
  cF = 1,
  crF = 0,
  ccF = cF,
  ccrF = crF,
  cM = 1,
  crM = 0,
  ccM = cM,
  ccrM = crM,
  dW = 0,
  drW = 0,
  ddW = dW,
  ddrW = drW,
  hSuf = 1,
  crossF = 0.5,
  crossM = 0.5,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)

Arguments

cF

Female cutting rate, one ClvR allele

crF

Female functional resistance rate, one ClvR allele

ccF

Female cutting rate, two ClvR alleles

ccrF

Female functional resistance rate, two ClvR alleles

cM

Male cutting rate, one ClvR allele

crM

Male functional resistance rate, one ClvR allele

ccM

Male cutting rate, two ClvR alleles

ccrM

Male functional resistance rate, two ClvR alleles

dW

Female deposition cutting rate

drW

Female deposition functional resistance rate

ddW

Female deposition (HH) cutting rate

ddrW

Female deposition (HH) functional resistance rate

hSuf

Haplosufficiency level, default is completely sufficient

crossF

Crossover rate in females, 0 is completely linked, 0.5 is unlinked, 1.0 is complete divergence

crossM

Crossover rate in males, 0 is completely linked, 0.5 is unlinked, 1.0 is complete divergence

eta

Genotype-specific mating fitness

phi

Genotype-specific sex ratio at emergence

omega

Genotype-specific multiplicative modifier of adult mortality

xiF

Genotype-specific female pupatory success

xiM

Genotype-specific male pupatory success

s

Genotype-specific fractional reduction(increase) in fertility

Value

Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele, and all genotype-specific parameters.

Details

Female deposition is implemented incorrectly. Right now, it is performed on male alleles prior to zygote formation - it should happen post-zygote formation. Since this construct doesn't have HDR, this should be fine.
Additionally, it is assumed that deposition requries loaded Cas9-RNP complexes from the mother, having Cas9 and no maternal gRNA, even in the presence of paternal gRNA, will not result in maternal deposition mediated cleavage.

Copy-number dependent rates are based on Cas9, not gRNA. The assumption is that RNA is easier to produce, and therefore won't limit cleavage by Cas9.