Inheritance Cube: Confinable Homing, X-Linked

This function creates an X-linked confinable homing construct, it has 5 alleles at the first locus and 4 alleles at the second. No crossovers or homing occurs into the y chromosome

W: Wild-type

H: Homing allele

A: Antidote allele

R: No-cost resistance allele

B: Detrimental resistance allele

Y: Male allele

Usage

cubeConfinableHomingX(
  cF = 1,
  chF = 0,
  crF = 0,
  dR = 0,
  dB = 0,
  crossF = 0,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)

Arguments

cF

Cutting efficiency of drive allele at locus 1 in females

chF

Homing efficiency of drive allele at locus 1 in females

crF

Resistance allele generation rate at locus 1 in females

dR

Background mutation rate from W and H into R allele in males and females

dB

Background mutation rate from A into B allele in males and females

crossF

Female crossover rate.

eta

Genotype-specific mating fitness

phi

Genotype-specific sex ratio at emergence

omega

Genotype-specific multiplicative modifier of adult mortality

xiF

Genotype-specific female pupatory success

xiM

Genotype-specific male pupatory success

s

Genotype-specific fractional reduction(increase) in fertility

Value

Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele, and all genotype-specific parameters.