Inheritance Cube: ECHACRX

This function creates an X-linked ECHACR construct, it has 5 alleles at the first locus and 4 alleles at the second.

W: Wild-type

H: Homing allele

E: Eraser allele

R: No-cost resistance allele

B: Detrimental resistance allele

cHW: Rate of homing from H, W -> H transition

cEH: Rate of homing from E, H -> E transition

cEW2: Rate of homing from E, W -> E transition

Usage

cubeECHACRX(
  cHW = 1,
  cEHW = 1,
  cEW1 = 1,
  cEW2 = 1,
  cEH = 1,
  chHW = 0,
  crHW = 0,
  chEHW = 0,
  crEHW = 0,
  ceEW1 = 0,
  crEW1 = 0,
  ceEW2 = 0,
  crEW2 = 0,
  ceEH = 0,
  crEH = 0,
  d1 = 0,
  d2 = 0,
  d3 = 0,
  dHW = 0,
  dEH = 0,
  dEW = 0,
  drHW = 0,
  drEH = 0,
  drEW = 0,
  crossF = 0,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)

Arguments

cHW

Cutting efficiency of drive allele at locus 1

cEHW

Cutting efficiency of drive allele, in the presence of ECHACR element, at locus 1

cEW1

Cutting efficiency of ECHACR element into W at locus 1

cEW2

Cutting efficiency of ECHACR element into W at locus 2

cEH

Cutting efficiency of ECHACR element into H

chHW

Homing efficiency of drive allele at locus 1

crHW

Resistance allele efficiency of drive allele at locus 1

chEHW

Homing efficiency of drive allele, in the presence of ECHACR element, at locus 1

crEHW

Resistance allele efficiency of drive allele, in the presence of ECHACR element, at locus 1

ceEW1

Homing efficiency of ECHACR element into W at locus 1

crEW1

Resistance allele efficiency of ECHACR element into W at locus 1

ceEW2

Homing efficiency of ECHACR element into W at locus 2

crEW2

Resistance allele efficiency of ECHACR element into W at locus 2

ceEH

Homing efficiency of ECHACR element into H

crEH

Resistance allele efficiency of ECHACR element into H

d1

Background mutation rate from W into R allele

d2

Background mutation rate from H into R allele

d3

Background mutation rate from E into R allele

dHW

Female H deposition rate against W

dEH

Female E deposition rate against H

dEW

Female E deposition rate against W

drHW

Female resistance generation rate, from H allele

drEH

Female resistance generation rate, from E allele

drEW

Female resistance generation rate, from E allele

crossF

Female crossover rate. 0 is fully linked, 0.5 is unlinked, 1 is negatively linked

eta

Genotype-specific mating fitness

phi

Genotype-specific sex ratio at emergence

omega

Genotype-specific multiplicative modifier of adult mortality

xiF

Genotype-specific female pupatory success

xiM

Genotype-specific male pupatory success

s

Genotype-specific fractional reduction(increase) in fertility

Value

Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele, and all genotype-specific parameters.

Details

This inheritance pattern corresponds to the Active Genetic Neutralizing Elements for Halting or Deleting Gene Drives publication.