Inheritance Cube: Immunizing Reversal/Basic Reversal

This function creates an Immunizing Reversal construct, it has 5 alleles at 1 locus

W: Wild-type

H: Homing allele

E: Eraser allele

R: No-cost resistance allele

B: Detrimental resistance allele

cubeImmunizingReversalMF(cHWM = 1, cHWF = 1, cEWM = 1, cEWF = 1,
  cEHM = 1, cEHF = 1, chHWM = 0, chHWF = 0, crHWM = 0,
  crHWF = 0, ceEWM = 0, ceEWF = 0, crEWM = 0, crEWF = 0,
  ceEHM = 0, ceEHF = 0, crEHM = 0, crEHF = 0, dHW = 0, dEW = 0,
  dEH = 0, dhHW = 0, drHW = 0, deEW = 0, drEW = 0, deEH = 0,
  drEH = 0, eta = NULL, phi = NULL, omega = NULL, xiF = NULL,
  xiM = NULL, s = NULL)

Arguments

cHWM
Cutting efficiency of H into W in males

cHWF
Cutting efficiency of H into W in females

cEWM
Cutting efficiency of E into W in males

cEWF
Cutting efficiency of E into W in females

cEHM
Cutting efficiency of E into H in males

cEHF
Cutting efficiency of E into H in females

chHWM
Homing efficiency of H into W in males

chHWF
Homing efficiency of H into W in females

crHWM
Resistance efficiency of H into W in males

crHWF
Resistance efficiency of H into W in females

ceEWM
Homing efficiency of E into W in males

ceEWF
Homing efficiency of E into W in females

crEWM
Resistance efficiency of E into W in males

crEWF
Resistance efficiency of E into W in females

ceEHM
Homing efficiency of E into H in males

ceEHF
Homing efficiency of E into H in females

crEHM
Resistance efficiency of E into H in males

crEHF
Resistance efficiency of E into H in females

dHW
Deposition cutting efficiency of H into W

dEW
Deposition cutting efficiency of E in to W

dEH
Deposition cutting efficiency of E into H

dhHW
Deposition homing efficiency of H into W

drHW
Deposition resistance efficiency of H into W

deEW
Deposition homing efficiency of E into W

drEW
Deposition resistance efficiency of E into W

deEH
Deposition homing efficiency of E into H

drEH
Deposition resistance efficiency of E into H

eta
Genotype-specific mating fitness

phi
Genotype-specific sex ratio at emergence

omega
Genotype-specific multiplicative modifier of adult mortality

xiF
Genotype-specific female pupatory success

xiM
Genotype-specific male pupatory success

s
Genotype-specific fractional reduction(increase) in fertility

Value

Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele, and all genotype-specific parameters.

Details

This is the general form for an immunizing reversal drive. If the EW terms are 0, then this simplifies to a basic reversal drive.This drive handles different male and female homing rates, and female deposition from each allele, signifying differential expression from an autosome.