Inheritance Cube: Immunizing Reversal

This function creates an Immunizing Reversal construct, it has 5 alleles at 1 locus

W: Wild-type

H: Homing allele

E: Eraser allele

R: No-cost resistance allele

B: Detrimental resistance allele

Usage

cubeImmunizingReversalX(
  cHW = 1,
  cEW = 1,
  cEH = 1,
  chHW = 0,
  crHW = 0,
  ceEW = 0,
  crEW = 0,
  ceEH = 0,
  crEH = 0,
  dHW = 0,
  dEW = 0,
  dEH = 0,
  dhHW = 0,
  drHW = 0,
  deEW = 0,
  drEW = 0,
  deEH = 0,
  drEH = 0,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)

Arguments

cHW

Cutting efficiency of H into W

cEW

Cutting efficiency of E into W

cEH

Cutting efficiency of E into H

chHW

Homing efficiency of H at W

crHW

Resistance generation efficiency of H at W

ceEW

Homing efficiency of E at W

crEW

Resistance generation efficiency of E at W

ceEH

Homing efficiency of E at H

crEH

Resistance efficiency of E at H

dHW

H deposition efficiency against W

dEW

E deposition efficiency against W

dEH

E deposition efficiency against H

dhHW

H deposition homing efficiency against W

drHW

H deposition resistance efficiency against W

deEW

E deposition homing efficiency against W

drEW

E deposition resistance efficiency against W

deEH

E deposition homing efficiency against H

drEH

E deposition resistance efficiency against H

eta

Genotype-specific mating fitness

phi

Genotype-specific sex ratio at emergence

omega

Genotype-specific multiplicative modifier of adult mortality

xiF

Genotype-specific female pupatory success

xiM

Genotype-specific male pupatory success

s

Genotype-specific fractional reduction(increase) in fertility

Value

Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele, and all genotype-specific parameters.

Details

This is the general form for an immunizing reversal drive. If the c_EW and d_EW, parameters are all 0, then this simplifies to a basic reversal drive. This drive represents an X-linked IR drive.