Inheritance Cube: tGDX

The trans-complementing Gene Drive (tGD) is a 1-locus, 2 target site drive. The first target site corresponds to the Cas protein, the second to an effector gene and the gRNAs. There are two sets of gRNAs, because each target site may have different cutting/homing/resistance rates, and each sex can have different rates for all of those things. Additionally, the parent that you receive your Cas from dictates its efficiency. Therefor, this construct has 6 alleles at the first locus and 5 alleles at the second.

Locus 1

W: Wild-type

P: Paternal Cas9

M: Maternal Cas9

R: Resistant allele 1

B: Resistant allele 2

Y: Y allele

Locus 2

W: Wild-type

G: gRNAs

R: Resistant 1

B: Resistant 2

Y: Y allele

cubeTGDX(
  cM1 = 0,
  cM2 = 0,
  cP1 = 0,
  cP2 = 0,
  hM1 = 0,
  hM2 = 0,
  hP1 = 0,
  hP2 = 0,
  rM1 = 0,
  rM2 = 0,
  rP1 = 0,
  rP2 = 0,
  crM = 0,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)

Arguments

cM1
Maternally inherited Cas9 cutting rate at locus 1

cM2
Maternally inherited Cas9 cutting rate at locus 2

cP1
Paternally inherited Cas9 cutting rate at locus 1

cP2
Paternally inherited Cas9 cutting rate at locus 2

hM1
Maternally inherited Cas9 homing efficiency at locus 1

hM2
Maternally inherited Cas9 homing efficiency at locus 2

hP1
Paternally inherited Cas9 homing efficiency at locus 1

hP2
Paternally inherited Cas9 homing efficiency at locus 2

rM1
Maternally inherited Cas9 resistance efficiency at locus 1

rM2
Maternally inherited Cas9 resistance efficiency at locus 2

rP1
Paternally inherited Cas9 resistance efficiency at locus 1

rP2
Paternally inherited Cas9 resistance efficiency at locus 2

crM
Maternal crossover rate, 0 is completely linked, 0.5 is unlinked, 1.0 is complete divergence

eta
Genotype-specific mating fitness

phi
Genotype-specific sex ratio at emergence

omega
Genotype-specific multiplicative modifier of adult mortality

xiF
Genotype-specific female pupatory success

xiM
Genotype-specific male pupatory success

s
Genotype-specific fractional reduction(increase) in fertility

Value

Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele, and all genotype-specific parameters.

Details

This drive corresponds to the transcomplementing gene drive developed by the Gantz and Bier lab.