Inheritance Cube: Y-Linked X-Shredder

This function creates an inheritance cube to model a Y-linked X-Shredder construct. This construct resides on the Y chromosome, and chops the X chromosome into many pieces during male spermatogenesis, destroying the X chromosome. Thus, males only produce Y gametes.
This drive has 5 alleles at 1 locus:

X: Wild-type X chromosome

R: X chromosome resistant to destruction by the shredder construct

Y: Wild-type Y chromosome

A: Attacking Y chromosome, a Y chromosome with the shredder construct

B: Broken Y chromosome, a Y chromosome with a defunct shredder construct

Usage

cubeXShredderY(
  cX = 1,
  crX = 0,
  cB = 0,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)

Arguments

cX

Rate of X shredding (default is 1, complete shredding)

crX

Rate of resistance chromosome generation (default is 0)

cB

Rate of shredder construct breakdown (default is 0)

eta

Genotype-specific mating fitness

phi

Genotype-specific sex ratio at emergence

omega

Genotype-specific multiplicative modifier of adult mortality

xiF

Genotype-specific female pupatory success

xiM

Genotype-specific male pupatory success

s

Genotype-specific fractional reduction(increase) in fertility

Value

Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele, and all genotype-specific parameters.