Deterministic Inter-Patch Migration

Deterministic model of interpatch migration from each patch. popFemale/popMale is retrieved from each patch using get_femalePop_Patch/get_malePop_Patch. Migration location is determined from the supplied matrices, private$migrationFemale or private$migrationMale. Final migration status is held in private$fMoveArray or private$mMoveMat.
Batch migration is not used in the deterministic model.

Usage

oneDay_Migration_Deterministic_Network()

Details

This function handles outbound and inbound migration. See MGDrivE-Model, 'Migration' section for more details on how inter-patch migration is handled.