Stochastic Inter-Patch Migration

Stochastic model of interpatch migration from each patch. popFemale/popMale is retrieved from each patch using get_femalePop_Patch/get_malePop_Patch. Migration location is determined from the supplied matrices, private$migrationFemale or private$migrationMale. Migration is modeled as a Multinomial process parameterized by migration location probabilities corresponding to each patch . Movement is sampled from rmultinom.
Batch migration begins as a rbinom sampled from private$migrationBatch$batchProbs.If there is batch migration, the location of migration is sampled uniformly (see sample), parameterized by private$migrationBatch$moveProbs. The amount of each sex that migrations is sampled from rbinom, parameterized by private$migrationBatch$sexProbs.

Usage

oneDay_Migration_Stochastic_Network()

Details

This function handles outbound and inbound migration. See MGDrivE-Model, 'Migration' section for more details on how inter-patch migration is handled.