Quantiles Function

Calculate the given quantiles of a matrix.

Usage

quantileC(Trials, Probs)

Arguments

Trials

Integer matrix to calculate quantiles over

Probs

Vector of quantiles

Value

Numeric Matrix

Details

This function calculates the given quantiles over the rows of an integer matrix. It uses method 8 of the stat::quantiles() function. It gives the same result, to numerical accuracy, and is designed to handle matrix input. It is only designed to work on integer matrices!