Retrieve Output

Read in output from directory. The resulting object will be a nested list; outermost nesting dimension indexes runID, within runID elements are split by sex and innermost nesting is over patches.

retrieveOutput(readDir, verbose = TRUE)

Arguments

readDir
directory where output was written to; must not end in path seperator

verbose
Chatty? Default is TRUE

Value

Nested List

Examples

# NOT RUN {
# Example assumes user has run and analyzed MGDrivE.
#  See vignette for examples of how to do that.

# set read directory
fPath <- "path/to/split/aggregated/output"

# read in data as nested lists
dataList <- retrieveOutput(readDir = fPath)
# }