Set Initial Population

This hidden function distributes the population at time 0 in the steady-state conformation. This involves finding the number of mosquitoes in each day of the aquatic stages, and then splitting adults into male and female. Each stage is appropriately split amongst the initial population genotypes (see parameterizeMGDrivE). It internally calls calcLarvalDist to determine the distribution of larvae before setting the eggs and pupa from that.

Usage

set_initialPopulation_Patch(
  adultEQ = adultEQ,
  larvalEQ = larvalEQ,
  adultRatioF = adultRatioF,
  adultRatioM = adultRatioM,
  larvalRatio = larvalRatio,
  timeAq = timeAq,
  muAq = muAq,
  alpha = alpha
)

Arguments

adultEQ

Equilibrium number of adults

larvalEQ

Equilibrium number of larvae

adultRatioF

Genotype specific ratio for adult females

adultRatioM

Genotype specific ratio for adult males

larvalRatio

Genotype specific ratio for larvae

timeAq

Time for each aquatic stage

muAq

Aquatic death rate

alpha

Density-dependent centering parameter