Network Class Definition

A Network class object stores all the information for a simulation on a defined landscape.

Network

Format

An R6Class generator object

Constructor

networkParameters: see Network.Parameters

driveCube: an inheritance cube, see MGDrivE-Cube for available cubes

patchReleases: see Release_basicRepeatedReleases for examples on how to set up release schedules

migrationMale: a stochastic matrix whose dimensions conform to the number of patches

migrationFemale: a stochastic matrix whose dimensions conform to the number of patches

directory: character string of output directory

Methods

get_moveVar: see get_moveVar_Network

get_timeAq: see get_timeAq_Network

get_thetaAq: see get_thetaAq_Network

get_windowSize: see get_windowSize_Network

get_beta: see get_beta_Network

get_muAd: see get_muAd_Network

get_rm: see get_rm_Network

get_AdPopEQ: see get_AdPopEQ_Network

get_g: see get_g_Network

get_Rm: see get_Rm_Network

get_muAq: see get_muAq_Network

get_alpha: see get_alpha_Network

get_Leq_Network: see get_Leq_Network

get_driveCube_genotype: see get_driveCube_genotype_Network

get_driveCube_index: see get_driveCube_index_Network

get_tau: see get_tau_Network

get_genotypesID: see get_genotypesID_Network

get_genotypesN: see get_genotypesN_Network

get_wildType: see get_wildType_Network

get_eta: see get_eta_Network

get_phi: see get_phi_Network

get_omega: see get_omega_Network

get_xiF: see get_xiF_Network

get_xiM: see get_xiM_Network

get_s: see get_s_Network

get_releaseType: see get_releaseType_Network

get_patch: see get_patch_Network

get_patches: see get_patches_Network

get_nPatch: see get_nPatch_Network

get_directory: see get_directory_Network

get_simTime: see get_simTime_Network

get_conADM: see get_conADM_Network

get_conAF1: see get_conAF1_Network

close_allConnections: see close_allConnections_Network

get_tNow: see get_tNow_Network

get_migrationMale: see get_migrationMale_Network

get_migrationMaleRow: see get_migrationMaleRow_Network

set_migrationMale: see set_migrationMale_Network

get_migrationFemale: see get_migrationFemale_Network

get_migrationFemaleRow: see get_migrationFemaleRow_Network

set_migrationFemale: see set_migrationFemale_Network

get_patchReleases: see get_patchReleases_Network

oneDay_Migration: see oneDay_Migration_Network

reset: see reset_Network

oneRun: see oneRun_Network

oneDay: see oneDay_Network

Fields

parameters: see Network.Parameters

patches: a list of Patch objects

nPatch: number of patches

simTime: maximum time of simulation

driveCube: an inheritance cube

tNow: current time of simulation (time starts at 2 because time 1 is the initial equilibrium state)

runID: an identifier for the current simulation run, useful for Monte Carlo simulation

directory: a character string of where to store output

conADM: a connection to write male population dynamics out to

conAF1: a connection to write female population dynamics out to

migrationMale: a stochastic matrix whose dimensions conform to the number of patches

migrationFemale: a stochastic matrix whose dimensions conform to the number of patches

patchReleases: a list of release schedules for each patch