Network Class Definition

A Network class object stores all the information for a simulation on a defined landscape.

Format

An R6Class generator object

Constructor

params: see parameterizeMGDrivE

driveCube: an inheritance cube

patchReleases: see basicRepeatedReleases for examples on how to set up release schedules

migrationMale: a stochastic matrix whose dimensions conform to the number of patches

migrationFemale: a stochastic matrix whose dimensions conform to the number of patches

migrationBatch: a list of batch migration parameters. SeebasicBatchMigration

directory: character string of output directory

verbose: Chatty? Default is TRUE

Methods

get_timeAq: see get_timeAq_Network

get_beta: see get_beta_Network

get_muAd: see get_muAd_Network

get_muAq: see get_muAq_Network

get_alpha: see get_alpha_Network

get_drivecubeindex: see get_drivecubeindex_Network

get_tau: see get_tau_Network

get_genotypesID: see get_genotypesID_Network

get_genotypesN: see get_genotypesN_Network

get_eta: see get_eta_Network

get_phi: see get_phi_Network

get_omega: see get_omega_Network

get_xiF: see get_xiF_Network

get_xiM: see get_xiM_Network

get_s: see get_s_Network

get_nPatch: see get_nPatch_Network

get_conADM: see get_conM_Network

get_conADF: see get_conF_Network

get_tNow: see get_tNow_Network

get_patchReleases: see get_patchReleases_Network

oneDay_Migration: see oneDay_Migration_Deterministic_Network or see oneDay_Migration_Stochastic_Network

reset: see reset_Network

oneDay: see oneDay_Network

oneRun: see oneRun_Network

multRun: see multRun_Network

Fields

parameters: see parameterizeMGDrivE

patches: a list of Patch objects

nPatch: number of patches

simTime: maximum time of simulation

sampTime: how often to write output, tNow %% sampTime

driveCube: an inheritance cube

tNow: current time of simulation (time starts at 2 because time 1 is the initial equilibrium state)

runID: an identifier for the current simulation run, useful for Monte Carlo simulation

directory: a character string of where to store output

conADM: a connection to write male population dynamics out to

conADF: a connection to write female population dynamics out to

migrationMale: a stochastic matrix whose dimensions conform to the number of patches

migrationFemale: a stochastic matrix whose dimensions conform to the number of patches

migrationBatch: list of items for batch migration in stochastic sim.

mMoveMat: holder object for male migration

fMoveArray: holder object for female migration

patchReleases: a list of release schedules for each patch

Examples

 if (FALSE) { # \dontrun{
 # There are no simple examples for this, so looking at the vignettes would be
 #  most useful.

 # Complete manual with examples, but none explored in depth.
 vignette("MGDrivE-Examples", package = "MGDrivE")

 # One example, explored in great detail. This is probably more helpful.
 vignette("MGDrivE-Run", package = "MGDrivE")

 } # }