Patch Class Definition

A Patch is a single well-mixed population that is the smallest unit of simulation for MGDrivE.

Patch

Format

An R6Class generator object

Constructor

patchID: integer ID of this patch

genotypesID: character vector of genotypes

simTime: maximum time of simulation

windowSize: necessary memory window size for model

EGGt0: initial egg population, \(L_{eq}\)

LARt0: initial larval population

PUPt0: initial pupae population

ADMt0: initial adult male population, \(Ad_{eq}\)

AF1t0: initial adult female population, \(Ad_{eq}\)

maleReleases: integer ID of this patch

femaleReleases: female release schedule for this patch, see basicRepeatedReleases

larvaeReleases: male release schedule for this patch, see basicRepeatedReleases

Methods

get_patchID: see get_patchID_Patch

get_AF1new: see get_AF1new_Patch

set_AF1new: see set_AF1new_Patch

get_ADMnew: see get_ADMnew_Patch

set_ADMnew: see set_ADMnew_Patch

accumulate_ADMnew: see accumulate_ADMnew_Patch

get_EGG: see get_EGG_Patch

get_LAR: see get_LAR_Patch

get_PUP: see get_PUP_Patch

get_ADM: see get_M_Patch

get_AF1: see get_F_Patch

get_EGGdly: see get_EGGdly_Patch

get_LARdly: see get_LARdly_Patch

get_PUPdly: see get_PUPdly_Patch

get_ADMdly: see get_ADMdly_Patch

get_AF1dly: see get_AF1dly_Patch

get_maleMigration: see get_maleMigration_Patch

get_femaleMigration: see get_femaleMigration_Patch

set_NetworkPointer: see set_NetworkPointer_Patch

get_NetworkPointer: see get_NetworkPointer_Patch

reset: see reset_Patch

oneDay_initOutput: see oneDay_initOutput_Patch

oneDay_writeOutput: see oneDay_writeOutput_Patch

oneDay_migrationIn: see oneDay_migrationIn_Patch

oneDay_maleReleases: see oneDay_maleReleases_Patch

oneDay_femaleReleases: see oneDay_femaleReleases_Patch

oneDay_PopDynamics: see oneDay_PopDynamics_Patch

oneDay_updatePopulation: see oneDay_updatePopulation_Patch

oneDay_calcLarvalDensityDependentFactor: see oneDay_calcLarvalDensityDependentFactor_Patch

oneDay_calcCumulativeLarvalDensityDependentFactor: see oneDay_calcCumulativeLarvalDensityDependentFactor_Patch

oneDay_calcCumulativePupaDensityDependentFactor: see oneDay_calcCumulativePupaDensityDependentFactor_Patch

oneDay_migrationOut: see oneDay_migrationOut_stochastic_Patch or oneDay_migrationOut_deterministic_Patch

oneDay_ovipositG1: see oneDay_ovipositG1_stochastic_Patch or oneDay_ovipositG1_deterministic_Patch

oneDay_larSurvival: see oneDay_larSurvival_stochastic_Patch or oneDay_larSurvival_deterministic_Patch

oneDay_hatchingFract: see oneDay_hatchingFract_stochastic_Patch or oneDay_hatchingFract_deterministic_Patch

oneDay_larHatching: see oneDay_larHatching_stochastic_Patch or oneDay_larHatching_deterministic_Patch

oneDay_eggsFract2: see oneDay_eggsFract2_stochastic_Patch or oneDay_eggsFract2_deterministic_Patch

oneDay_larPupating: see oneDay_larPupating_stochastic_Patch or oneDay_larPupating_deterministic_Patch

oneDay_numMaleFemale: see oneDay_numMaleFemale_stochastic_Patch or oneDay_numMaleFemale_deterministic_Patch

oneDay_admSurvival: see oneDay_admSurvival_stochastic_Patch or oneDay_admSurvival_deterministic_Patch

oneDay_admPupating: see oneDay_admPupating_stochastic_Patch or oneDay_admPupating_deterministic_Patch

oneDay_af1Survival: see oneDay_af1Survival_stochastic_Patch or oneDay_af1Survival_deterministic_Patch

oneDay_af1Pupation: see oneDay_af1Pupation_stochastic_Patch or oneDay_af1Pupation_deterministic_Patch

oneDay_af1Mating: see oneDay_af1Mating_stochastic_Patch or oneDay_af1Mating_deterministic_Patch

Fields

patchID: integer ID of this patch

EGGt0: vector of initial egg stage population

LARt0: vector of initial larval stage population

PUPt0: vector of initial pupae stage population

ADMt0: vector of initial adult male stage population

AF1t0: matrix of initial adult female stage population

EGG: egg stage population

LAR: larvae stage population

PUP: pupae stage population

ADM: adult male stage population

AF1: adult female stage population

EGGdly: delay egg stage population

LARdly: delay larvae stage population

PUPdly: delay pupae stage population

ADMdly: delay adult male stage population

AF1dly: delay adult female stage population

LARnew: new larval population after difference equations; needed to store population prior to migration exchange

ADMnew: new adult male population after difference equations; needed to store population prior to migration exchange

AF1new: new adult female population after difference equations; needed to store population prior to migration exchange

maleMigration: matrix of outbound migrating males of dimension nGenotypes X nPatch

femaleMigration: array of outbound migrating females of dimension nGenotypes X nGenotypes X nPatch

NetworkPointer: a reference to enclosing Network

ovipositG1: new eggs after oviposition by mated female mosquitoes

larSurvival: surviving larvae

hatchingFract: fraction of larvae that hatch

larPupating: fraction of larvae that undergo pupation

numMaleFemale: number of male vs. female emerging imago stage adults

admSurvival: number of surviving adult males

admPupating: number of pupating imago stage adults that become males

af1Survival: number of surviving adult females

af1Pupation: number of pupating imago stage adults that become females

maleMatrix: row of male migration matrix corresponding to migration from this patch

femaleMatrix: row of female migration matrix corresponding to migration from this patch

larDDMortal: larval mortality

f: density dependent factor in larval mortality