Make List of Modified Mosquito Releases

Sets up a release schedule for a single patch, returns a list to be used in oneDay_maleReleases_Patch or oneDay_femaleReleases_Patch.

basicRepeatedReleases(genotypes, releaseStart, releaseEnd, releaseInterval,
  releaseVector, sex = "M")

Arguments

genotypes
possible genotypes

releaseStart
day releases start

releaseEnd
day releases end

releaseInterval
interval between releases

releaseVector
named character vector of release composition

sex
character in 'M','F','E'

Examples

# to setup for 3 patches but only release in the first with a defined release schedule:

patchReleases = replicate(n = 3,expr = {
  list(maleReleases = NULL,femaleReleases = NULL)
},simplify = FALSE)

patchReleases[[1]]$femaleReleases = basicRepeatedReleases(genotypes = cubeHoming1RA$genotypesID,
                                                          releaseStart = 5,
                                                          releaseEnd = 30,
                                                          releaseInterval = 5,
                                                          releaseVector = c("HH"=100,
                                                                            "Hh"=0,
                                                                            "HR"=0,
                                                                            "hh"=0,
                                                                            "hR"=0,
                                                                            "RR"=0),
                                                          sex = "F")
patchReleases[[1]]$maleReleases = basicRepeatedReleases(genotypes = cubeHoming1RA$genotypesID,
                                                        releaseStart = 50,
                                                        releaseEnd = 60,
                                                        releaseInterval = 1,
                                                        releaseVector = c("HH"=100,
                                                                          "Hh"=0,
                                                                          "HR"=0,
                                                                          "hh"=0,
                                                                          "hR"=0,
                                                                          "RR"=0),
                                                        sex = "M")