Inheritance Cube: ECHACRX

This function creates an X-linked ECHACR construct, it has 5 alleles at the first locus and 4 alleles at the second.

W: Wild-type

H: Homing allele

E: Eraser allele

R: No-cost resistance allele

B: Detrimental resistance allele

cHW: Rate of homing from H, W -> H transition

cEH: Rate of homing from E, H -> E transition

cEW: Rate of homing from E, W -> E transition

cubeECHACRX(cHW = 1, cEW = 1, cEH = 1, chHW = 0, crHW = 0,
  ceEW = 0, crEW = 0, ceEH = 0, crEH = 0, d1 = 0, d2 = 0,
  d3 = 0, dHW = 0, dEH = 0, dEW = 0, drHW = 0, drEH = 0,
  drEW = 0, eta = NULL, phi = NULL, omega = NULL, xiF = NULL,
  xiM = NULL, s = NULL)

Arguments

cHW
Cutting efficiency of drive allele at locus 1

cEW
Cutting efficiency of ECHACR element into W

cEH
Cutting efficiency of ECHACR element into H

chHW
Homing efficiency of drive allele at locus 1

crHW
Resistance allele efficiency of drive allele at locus 1

ceEW
Homing efficiency of ECHACR element into W

crEW
Resistance allele efficiency of ECHACR element into W

ceEH
Homing efficiency of ECHACR element into H

crEH
Resistance allele efficiency of ECHACR element into H

d1
Background mutation rate from W into R allele

d2
Background mutation rate from H into R allele

d3
Background mutation rate from E into R allele

dHW
Female H deposition rate against W

dEH
Female E deposition rate against H

dEW
Female E deposition rate against W

drHW
Female resistance generation rate, from H allele

drEH
Female resistance generation rate, from E allele

drEW
Female resistance generation rate, from E allele

eta
Genotype-specific mating fitness

phi
Genotype-specific sex ratio at emergence

omega
Genotype-specific multiplicative modifier of adult mortality

xiF
Genotype-specific female pupatory success

xiM
Genotype-specific male pupatory success

s
Genotype-specific fractional reduction(increase) in fertility

Value

Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele, and all genotype-specific parameters.