Inheritance Cube: Split CRISPR Drive with 2 Resistance Alleles and male/female specific homing

This is a sex-specific version of a split CRISPR drive. At one locus is the Cas9, inherited in a Mendelian fashion. At a second, unlinked, locus are the gRNAs. When the two loci occur together, the gRNAs drive, with potential damaged alleles, but the Cas9 remains Mendelian. It is assumed that this is an autosomal drive. This drive corresponds to the confinable gene drive system developed by the Akbari lab.

Usage

cubeSplitDrive(
  cM = 1,
  chM = 0,
  crM = 0,
  ccM = cM,
  cchM = chM,
  ccrM = crM,
  cF = 1,
  chF = 0,
  crF = 0,
  ccF = cF,
  cchF = chF,
  ccrF = crF,
  dW = 0,
  dhW = 0,
  drW = 0,
  ddW = dW,
  ddhW = dhW,
  ddrW = drW,
  eta = NULL,
  phi = NULL,
  omega = NULL,
  xiF = NULL,
  xiM = NULL,
  s = NULL
)

Arguments

cM

Cutting efficiency in males, one Cas9 allele

chM

Homing efficiency in males, one Cas9 allele

crM

Resistance efficiency in males, one Cas9 allele

ccM

Cutting efficiency in males, two Cas9 alleles

cchM

Homing efficiency in males, two Cas9 alleles

ccrM

Resistance efficiency in males, two Cas9 alleles

cF

Cutting efficiency in females, one Cas9 allele

chF

Homing efficiency in females, one Cas9 allele

crF

Resistance efficiency in females, one Cas9 allele

ccF

Cutting efficiency in females, two Cas9 alleles

cchF

Homing efficiency in females, two Cas9 alleles

ccrF

Resistance efficiency in females, two Cas9 alleles

dW

Maternal deposition cutting, one Cas9 allele

dhW

Maternal deposition homing, one Cas9 allele

drW

Maternal deposition resistance, one Cas9 allele

ddW

Maternal deposition cutting, two Cas9 alleles

ddhW

Maternal deposition homing, two Cas9 alleles

ddrW

Maternal deposition resistance, two Cas9 alleles

eta

Genotype-specific mating fitness

phi

Genotype-specific sex ratio at emergence

omega

Genotype-specific multiplicative modifier of adult mortality

xiF

Genotype-specific female pupatory success

xiM

Genotype-specific male pupatory success

s

Genotype-specific fractional reduction(increase) in fertility

Value

Named list containing the inheritance cube, transition matrix, genotypes, wild-type allele, and all genotype-specific parameters.