parameterizeMGDrivE

Generate parameters for simulation on a Network. Parameters include: average generation time \(g\), population growth rate \(R_{m}\), aquatic mortality \(\mu_{Aq}\), and aquatic survival \(\theta_{Aq}\), which are shared between patches and calculated by calcAverageGenerationTime, calcPopulationGrowthRate, and calcLarvalStageMortalityRate.
Patch-specific parameters \(\alpha\) and \(L_{eq}\) are calculated for each patch by calcDensityDependentDeathRate and calcLarvalPopEquilibrium.

parameterizeMGDrivE(
  runID = 1L,
  nPatch,
  simTime,
  sampTime = 1L,
  tEgg = 1L,
  tLarva = 14L,
  tPupa = 1L,
  beta = 32,
  muAd = 0.123,
  popGrowth = 1.096,
  AdPopEQ,
  LarPopRatio,
  AdPopRatio_F,
  AdPopRatio_M,
  inheritanceCube
)

Arguments

runID
Begin counting runs with this set of parameters from this value

nPatch
Number of Patch

simTime
Maximum time to run simulation

sampTime
Times to sample, used as tNow %% sampTime, default is every day

tEgg
Length of egg stage

tLarva
Length of larval instar stage

tPupa
Length of pupal stage

beta
Female egg batch size of wild-type

muAd
Wild-type daily adult mortality (1/muAd is average wild-type lifespan)

popGrowth
Daily population growth rate (used to calculate equilibrium)

AdPopEQ
Single number or vector of adult population size at equilibrium (single number implies all patches have the same population)

LarPopRatio
May be empty; if not, a vector gives the wildtype gene frequencies among larval stages at the beginning of simulation or a matrix provides different initial frequencies for each patch (every row is a different patch, must have nrow = nPatch)

AdPopRatio_F
May be empty; if not, a vector gives the wildtype gene frequencies among adult females at the beginning of simulation or a matrix provides different initial frequencies for each patch (every row is a different patch, must have nrow = nPatch)

AdPopRatio_M
May be empty; if not, a vector gives the wildtype gene frequencies among adult males at the beginning of simulation or a matrix provides different initial frequencies for each patch (every row is a different patch, must have nrow = nPatch)

inheritanceCube
Inheritance cube to check/set population ratios at the beginning of the simulation

Examples

# using default parameters for 2 patches
#  using different population sizes for patches
simPars <- parameterizeMGDrivE(nPatch = 2, simTime = 365,
                               AdPopEQ = c(100,200), inheritanceCube = cubeMendelian())