Plot

Plots several traces from MGDrive, assuming each set is another repetition from the same experiment.
Given the readDirectory, this function assumes the follow file structure:

readDirectory

repetition 1

patch 1

patch 2

patch 3

repetition 2

patch 1

patch 2

patch 3

repetition 3

repetition 4

...

plotMGDrivEMult(readDir, whichPatches = NULL, totalPop = FALSE,
                       nonZeroGen = FALSE, lwd = 0.75, alpha = 0.75)

Arguments

readDir
Directory to find repetition folders in

whichPatches
Vector of patches to plot, must be less than 15. Default is NULL if less than 15 patches

totalPop
Boolean, to plot the total population or not. Default is FALSE

nonZeroGen
Boolean, to plot genotypes that are always zero in simulation

lwd
Double, specify the line width for plotting

alpha
Double, specify the opacity for plotting

Details

This function plots output from one run or one set of runs after being analyzed. Setting totalPop to FALSE keeps it from plotting the total population. NonZeroGen accounts for genotypes that could exist, but are not created in the simulation. Default is FALSE, as this is easier to read on a plot.

Examples

# Requires the user to have run MGDrivE with repetitions and keep them in the
#  directory organization shown above.
# See vignette for complete example