Plot

Plots several traces from MGDrivE, assuming each set is another repetition from the same experiment.
Given the readDir, this function assumes the follow file structure:

readDir

repetition 1

patch 1

patch 2

patch 3

repetition 2

patch 1

patch 2

patch 3

repetition 3

repetition 4

...

plotMGDrivEMult(readDir, whichPatches = NULL, totalPop = FALSE,
                       nonZeroGen = FALSE, lwd = 0.75, alpha = 0.75)

Arguments

readDir
Directory to find repetition folders in

whichPatches
Vector of patches to plot, must be less than 15. Default is NULL if less than 15 patches

totalPop
Boolean, to plot the total population or not. Default is FALSE

nonZeroGen
Boolean, to plot genotypes that are always zero in simulation

lwd
Double, specify the line width for plotting

alpha
Double, specify the opacity for plotting

Details

This function plots output from one run or one set of runs after being analyzed. Setting totalPop to FALSE keeps it from plotting the total population. NonZeroGen accounts for genotypes that could exist, but are not created in the simulation. Default is FALSE, as this is easier to read on a plot.

Examples

if (FALSE) {
# Requires the user to have run MGDrivE, logically stochastic, analyzed
#  the data, and stored it in the directory shown below.
# See vignette for complete example

# Folder where single run is stored
fPath <- "path/to/data/containing/folder"

# plot output to see effect
plotMGDrivEMult(readDir=fPath,totalPop = TRUE,lwd=3.5,alpha=1)
}